+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - New constructor methods `logHintsMultiline` to print logs with break line instead of new line
+- New generator `view` to create a new view in the views folder
 
 ### Changed
 - Removed npm debug log
@@ -16,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Refactored the prompt steps to assure consistency of hints, new lines and syntax
 - Moved the utils files (colors, utils) in the common folder
 - Removed the constructor and exported all the logs utils functions into a new logs module in common folder
+- Moved the partial folder in the common folder
 
 ### Fixed
 - Wrong theme name `cozen-portail-france` replaced by `altran-portail-france`